Added binary search for the number of digits
and cut down the variable "car" and replaced with variable "j".
In the binary search there is extra if (line 11) because most of the
time i will be smaller than j by one, during the while.(ex i=2 j=3)and
because i and j are both int (3+2/2 = 5/2 = 2) means the i wont be able
to have the value 3,ever, so the while keeps running forever. But the
extra else(line 12) does that, assigning i to j when i is smaller than j
by one(line 12).